+++ lib/models/data/all_team_data.dart
@@ -1,8 +1,8 @@
 import "package:scouting_frontend/models/enums/climb_enum.dart";
-import "package:scouting_frontend/models/team_data/pit_data/pit_data.dart";
-import "package:scouting_frontend/models/team_data/technical_match_data.dart";
+import "package:scouting_frontend/models/data/pit_data/pit_data.dart";
+import "package:scouting_frontend/models/data/technical_match_data.dart";
 import "package:scouting_frontend/models/team_model.dart";
-import "package:scouting_frontend/models/team_data/aggregate_data/aggregate_technical_data.dart";
+import "package:scouting_frontend/models/data/aggregate_data/aggregate_technical_data.dart";
 import "package:scouting_frontend/models/enums/robot_field_status.dart";
 
 class AllTeamData {
