+++ lib/models/data/technical_data.dart
