+++ lib/models/data/technical_match_data.dart
@@ -1,7 +1,7 @@
 import "package:scouting_frontend/models/enums/climb_enum.dart";
 import "package:scouting_frontend/models/match_identifier.dart";
-import "package:scouting_frontend/models/team_data/starting_position_enum.dart";
-import "package:scouting_frontend/models/team_data/technical_data.dart";
+import "package:scouting_frontend/models/data/starting_position_enum.dart";
+import "package:scouting_frontend/models/data/technical_data.dart";
 import "package:scouting_frontend/models/enums/robot_field_status.dart";
 
 class TechnicalMatchData {
