+++ lib/models/fetch_functions/fetch_single_team.dart
@@ -1,7 +1,7 @@
 import "dart:collection";
 import "package:flutter/material.dart";
 import "package:scouting_frontend/models/fetch_functions/fetch_teams.dart";
-import "package:scouting_frontend/models/team_data/team_data.dart";
+import "package:scouting_frontend/models/data/team_data.dart";
 
 Future<TeamData> fetchSingleTeamData(
   final int id,
