+++ lib/models/team_info_models/quick_data.dart
@@ -1,4 +1,4 @@
-import "package:scouting_frontend/models/team_data/technical_data.dart";
+import "package:scouting_frontend/models/data/technical_data.dart";
 
 class QuickData {
   QuickData({
