+++ lib/views/pc/alliance_auto_planner/auto_planner_screen.dart
@@ -1,8 +1,8 @@
 import "package:collection/collection.dart";
 import "package:flutter/material.dart";
 import "package:scouting_frontend/models/id_providers.dart";
-import "package:scouting_frontend/models/team_data/starting_position_enum.dart";
-import "package:scouting_frontend/models/team_data/team_data.dart";
+import "package:scouting_frontend/models/data/starting_position_enum.dart";
+import "package:scouting_frontend/models/data/team_data.dart";
 import "dart:ui" as ui;
 import "package:scouting_frontend/models/team_model.dart";
 import "package:scouting_frontend/views/common/auto_path/multiple_path_canvas.dart";
