+++ lib/views/pc/compare/compare_screen.dart
@@ -5,7 +5,7 @@ import "package:orbit_standard_library/orbit_standard_library.dart";
 import "package:scouting_frontend/models/id_providers.dart";
 import "package:scouting_frontend/models/team_model.dart";
 import "package:scouting_frontend/models/fetch_functions/fetch_teams.dart";
-import "package:scouting_frontend/models/team_data/team_data.dart";
+import "package:scouting_frontend/models/data/team_data.dart";
 import "package:scouting_frontend/views/common/team_selection_future.dart";
 import "package:scouting_frontend/views/constants.dart";
 import "package:scouting_frontend/views/common/card.dart";
