+++ lib/views/pc/compare/widgets/lineChart/compare_line_chart.dart
@@ -2,8 +2,8 @@ import "dart:math";
 import "package:flutter/material.dart";
 import "package:scouting_frontend/models/enums/match_type_enum.dart";
 import "package:scouting_frontend/models/match_identifier.dart";
-import "package:scouting_frontend/models/team_data/team_data.dart";
-import "package:scouting_frontend/models/team_data/technical_match_data.dart";
+import "package:scouting_frontend/models/data/team_data.dart";
+import "package:scouting_frontend/models/data/technical_match_data.dart";
 import "package:scouting_frontend/views/common/dashboard_linechart.dart";
 
 class CompareLineChart extends StatelessWidget {
