+++ lib/views/pc/picklist/picklist_card.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 import "package:scouting_frontend/views/common/card.dart";
-import "package:scouting_frontend/models/team_data/all_team_data.dart";
+import "package:scouting_frontend/models/data/all_team_data.dart";
 import "package:scouting_frontend/views/constants.dart";
 import "package:scouting_frontend/views/pc/picklist/auto_picklist_popup.dart";
 import "package:scouting_frontend/views/pc/picklist/pick_list_screen.dart";
